# backend/docs/swagger.yaml
basePath: /
definitions:
...
  gin.H:
    additionalProperties: {}
    type: object
...
  models.Group:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@servasec.local
    name: API Support
  description: Security scanner management platform API
  title: Servasec API
  version: "1.0"
paths:
...
  /api/ingest:
    post:
      consumes:
      - multipart/form-data
      description: Upload a scan results file. Authenticate via API token header,
        application ID in path, or application slug in path.
      parameters:
      - description: Version name
        in: formData
        name: version
        type: string
      - description: Branch name
        in: formData
        name: branch
        type: string
      - description: Scanner type name
        in: formData
        name: scannerType
        type: string
      - description: Scan results file (SARIF, etc.)
        in: formData
        name: file
        required: true
        type: file
      - description: API token (alternative to id/slug auth)
        in: header
        name: X-Api-Token
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Scan created with findings count
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Application not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Ingest scan results
      tags:
      - Scans
...
  /applications/{id}/ingest:
    post:
      consumes:
      - multipart/form-data
      description: Upload a scan results file. Authenticate via API token header,
        application ID in path, or application slug in path.
      parameters:
      - description: Application ID
        in: path
        name: id
        type: string
      - description: Version name
        in: formData
        name: version
        type: string
      - description: Branch name
        in: formData
        name: branch
        type: string
      - description: Scanner type name
        in: formData
        name: scannerType
        type: string
      - description: Scan results file (SARIF, etc.)
        in: formData
        name: file
        required: true
        type: file
      - description: API token (alternative to id/slug auth)
        in: header
        name: X-Api-Token
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Scan created with findings count
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Application not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Ingest scan results
      tags:
      - Scans
...
  /groups:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: List of groups
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
      summary: List groups
      tags:
      - Groups
    post:
      consumes:
      - application/json
      parameters:
      - description: Group details
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created group
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create group
      tags:
      - Groups
  /groups/{id}:
    delete:
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Group deleted
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete group
      tags:
      - Groups
    get:
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Group details
          schema:
            $ref: '#/definitions/models.Group'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get group by ID
      tags:
      - Groups
    put:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated group
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update group
      tags:
      - Groups
...
securityDefinitions:
  ApiKeyAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"